+++ schemas/track.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Track(BaseModel):
+    title: str
+    author: str
+    thumbnail_url: str
