+++ schemas/user.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class User(BaseModel):
+    username: str
+    password: str
+
+
+class ShowUser(BaseModel):
+    key: str
+    username: str
